+++ Runtime/Internal/PropertyBagDefinitions/LightPropertyBagDefinition.cs
@@ -5,7 +5,7 @@ namespace Unity.RuntimeSceneSerialization.Internal
 {
     static class LightPropertyBagDefinition
     {
-        [RuntimeInitializeOnLoadMethod]
+        [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSceneLoad)]
 #if UNITY_EDITOR
         [UnityEditor.InitializeOnLoadMethod]
 #endif