+++ Runtime/Internal/PropertyBagDefinitions/RigidbodyPropertyBagDefinition.cs
@@ -6,7 +6,7 @@ namespace Unity.RuntimeSceneSerialization.Internal
 {
     static class RigidbodyPropertyBagDefinition
     {
-        [RuntimeInitializeOnLoadMethod]
+        [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSceneLoad)]
 #if UNITY_EDITOR
         [UnityEditor.InitializeOnLoadMethod]
 #endif